reorganize document structure
