+++ packages/ts/oa42-generator/src/documents/oas30/index.ts
@@ -1,3 +1,2 @@
 export * from "./document.js";
 export * from "./factory.js";
-export * from "./selectors.js";
