+++ packages/ts/oa42-generator/src/documents/swagger2/document.ts
@@ -3,7 +3,10 @@ import * as models from "../../models/index.js";
 import { DocumentBase } from "../document-base.js";
 
 export class Document extends DocumentBase<oas.SchemaJson> {
-  public getApiModel(): Promise<models.Api> {
+  public getApiModel(): models.Api {
+    throw new Error("Method not implemented.");
+  }
+  protected selectSchemas(pointer: string, document: oas.SchemaJson): Iterable<[string, unknown]> {
     throw new Error("Method not implemented.");
   }
 }
